feat: Add a script that prepare postgres database
Responsible for all database preparation and empty table creation.
The script does the following:
- create new database "swile"
- create new schema "general"
- create source tables

It has a mix of DBT run operations with macros as well as postgres connector execution commands as I wanted to test the connection to the database first before I check the dbt project and did with the database connection library. It can be replaced later with DBT pre-hooks