feat: Add pipline logic (#2)
* feat: Add a constants.py file that contain logic configs

* feat: Add a script with logic to read and write transactions

The script contains functions that will be used within an airflow dag to do the following:
- read json files from MinIO
- write json data into a postgresql server

* feat: Add a script that prepare postgres database

Responsible for all database preparation and empty table creation.
The script does the following:
- create new database "swile"
- create new schema "general"
- create source tables

It has a mix of DBT run operations with macros as well as postgres connector execution commands as I wanted to test the connection to the database first before I check the dbt project and did with the database connection library. It can be replaced later with DBT pre-hooks

* feat: Add a script that send api requests to get missing naf_codes

The script contain the logic that will be used within the airflow dag

* feat: Add an airflow dag for the automated pipeline process

This DAG is responsible for the entire automated process of from read json data from data storage to generating the final report table "spent".

* feat: Add an initial requirement.txt file

This file was generated from own Windows machine environment which might contain libraries that aren't available on linux or different dependencies. It will be replaced with a shorter version with only needed libraries later on.