fix: Replace ORM operation with direct database connector ones
Due to a conflict in SQLAlchemy version the default pandas method to write and read dataframes using ORM is broken and thus had to revert back to a method that doesn't use SQLAlchemy.
The conflict is caused by flask-appBuilder (airflow dependancy) requiring a version of SQLAlchemy incompatible with pandas and psycopg2.